+++ packages/nimble-components/src/anchor-menu-item/styles.ts
@@ -12,17 +12,13 @@ import {
     iconColor,
     iconSize
 } from '../theme-provider/design-tokens';
+import { focusVisible } from '../utilities/style/focus';
 
 export const styles = css`
     ${display('grid')}
 
     :host {
         font: ${bodyFont};
-    }
-
-    :host(:focus-within:not(:active)) {
-        outline: 2px solid ${borderHoverColor};
-        outline-offset: -2px;
     }
 
     :host(:hover) {
@@ -54,6 +50,13 @@ export const styles = css`
         cursor: pointer;
         text-decoration: none;
         outline: none;
+        padding-left: 8px;
+        padding-right: 8px;
+    }
+
+    a${focusVisible} {
+        outline: 2px solid ${borderHoverColor};
+        outline-offset: -2px;
     }
 
     :host([disabled]) a {