+++ packages/nimble-components/src/mapping/spinner/tests/mapping-spinner.react.tsx
@@ -1,4 +1,0 @@
-import { MappingSpinner } from '..';
-import { wrap } from '../../../utilities/tests/react-wrapper';
-
-export const NimbleMappingSpinner = wrap(MappingSpinner);