+++ packages/nimble-components/src/patterns/dropdown/styles.ts
@@ -9,6 +9,7 @@ import {
     borderWidth,
     controlHeight,
     iconSize,
+    menuMinWidth,
     popupBorderColor,
     smallDelay,
     smallPadding,
@@ -36,7 +37,7 @@ export const styles = css`
         position: relative;
         justify-content: center;
         ${userSelectNone}
-        min-width: 250px;
+        min-width: ${menuMinWidth};
         outline: none;
         vertical-align: top;
         --ni-private-hover-indicator-width: calc(${borderWidth} + 1px);