+++ packages/nimble-components/src/table-column/icon/tests/table-column-icon.react.tsx
@@ -1,4 +1,0 @@
-import { TableColumnIcon } from '..';
-import { wrap } from '../../../utilities/tests/react-wrapper';
-
-export const NimbleTableColumnIcon = wrap(TableColumnIcon);