+++ packages/nimble-components/src/table-column/number-text/tests/table-column-number-text.react.tsx
@@ -1,4 +1,0 @@
-import { TableColumnNumberText } from '..';
-import { wrap } from '../../../utilities/tests/react-wrapper';
-
-export const NimbleTableColumnNumberText = wrap(TableColumnNumberText);