+++ packages/nimble-components/src/text-area/tests/text-area.react.tsx
@@ -1,4 +1,0 @@
-import { TextArea } from '..';
-import { wrap } from '../../utilities/tests/react-wrapper';
-
-export const NimbleTextArea = wrap(TextArea);