+++ packages/storybook/src/nimble/button/button.react.tsx
@@ -1,0 +1,4 @@
+import { Button } from '@ni/nimble-components/dist/esm/button';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleButton = wrap(Button);