+++ packages/storybook/src/nimble/mapping/icon/mapping-icon.react.tsx
@@ -1,0 +1,4 @@
+import { MappingIcon } from '@ni/nimble-components/dist/esm/mapping/icon';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleMappingIcon = wrap(MappingIcon);