+++ packages/storybook/src/nimble/rich-text/editor/rich-text-editor.react.tsx
@@ -1,0 +1,4 @@
+import { RichTextEditor } from '@ni/nimble-components/dist/esm/rich-text/editor';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleRichTextEditor = wrap(RichTextEditor);