+++ packages/storybook/src/nimble/tab-panel/tab-panel.react.tsx
@@ -1,0 +1,4 @@
+import { TabPanel } from '@ni/nimble-components/dist/esm/tab-panel';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleTabPanel = wrap(TabPanel);