+++ packages/storybook/src/nimble/table-column/duration-text/table-column-duration-text.react.tsx
@@ -1,0 +1,4 @@
+import { TableColumnDurationText } from '@ni/nimble-components/dist/esm/table-column/duration-text';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleTableColumnDurationText = wrap(TableColumnDurationText);