+++ packages/storybook/src/nimble/table-column/number-text/table-column-number-text.react.tsx
@@ -1,0 +1,4 @@
+import { TableColumnNumberText } from '@ni/nimble-components/dist/esm/table-column/number-text';
+import { wrap } from '../../../utilities/react-wrapper';
+
+export const NimbleTableColumnNumberText = wrap(TableColumnNumberText);