+++ packages/storybook/src/nimble/tabs/tabs.react.tsx
@@ -1,0 +1,4 @@
+import { Tabs } from '@ni/nimble-components/dist/esm/tabs';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleTabs = wrap(Tabs);