+++ packages/storybook/src/nimble/toolbar/toolbar.react.tsx
@@ -1,0 +1,4 @@
+import { Toolbar } from '@ni/nimble-components/dist/esm/toolbar';
+import { wrap } from '../../utilities/react-wrapper';
+
+export const NimbleToolbar = wrap(Toolbar);